Distribution.HaskellSuite.Modules

Module monad

When you need to resolve modules, you work in a ModuleT monad (or another monad that is an instance of MonadModule) and use the getModuleInfo function.

It finds an installed module by its name and reads (and caches) its info from the info file. Then you run a ModuleT monadic action using evalModuleT or runModuleT.

To run a ModuleT action you'll also need to provide the set of packages (represented by their InstalledPackageInfo) in which to search for modules. You can get such a set from either getInstalledPackages or readPackagesInfo, depending on your use case.

data ModuleT i m a

A standard module monad transformer.

i is the type of module info, m is the underlying monad.

getModuleInfo :: (MonadModule m, ModName n) => n -> m (Maybe (ModuleInfo m))

Tries to find the module in the current set of packages, then find the module's info file, and reads and caches its contents.

Returns Nothing if the module could not be found in the current set of packages. If the module is found, but something else goes wrong (e.g. there's no info file for it), an exception is thrown.

evalModuleT

Arguments

:: MonadIO m

=> ModuleT i m a
the monadic action to run

-> Packages
packages in which to look for modules

-> String
file extension of info files

-> (FilePath -> m i)
how to read information from an info file

-> m a

Run a ModuleT action.

This is a simplified version of runModuleT.

runModuleT

Arguments

:: MonadIO m

=> ModuleT i m a
the monadic action to run

-> Packages
packages in which to look for modules

-> String
file extension of info files

-> (FilePath -> m i)
how to read information from an info file

-> Map ModuleName i
initial set of module infos

-> m (a, Map ModuleName i)
return value, plus all cached module infos (that is, the initial set plus all infos that have been read by the action itself)

Run a ModuleT action

class Monad m => MonadModule m where

This class defines the interface that is used by getModuleInfo, so that you can use it in monads other than ModuleT.

You don't typically have to define your own instances of this class, but here are a couple of cases when you might:

A pure (non-MonadIO) mockup module monad for testing purposes

A transformer over ModuleT

You need a more complex way to retrieve the module info

Associated Types

type ModuleInfo m

The type of module info

Methods

lookupInCache :: ModName n => n -> m (Maybe (ModuleInfo m))

insertInCache :: ModName n => n -> ModuleInfo m -> m ()

getPackages :: m Packages

readModuleInfo :: ModName n => [FilePath] -> n -> m (ModuleInfo m)

Read the module info, given a list of search paths and the module name

Module names

class ModName n where

Different libraries (Cabal, haskell-src-exts, ...) use different types to represent module names. Hence this class.

Methods

modToString :: n -> String

convertModuleName :: ModName n => n -> ModuleName

Convert module name from arbitrary representation to Cabal's one